feat: changes to handle display_name field in endpoint return and other minor fixes

+++ internal/kafka/internal/presenters/kafka.go
@@ -83,10 +83,10 @@ func PresentKafkaRequest(kafkaRequest *dbapi.KafkaRequest, config *config.KafkaC
 		InstanceTypeName:            displayName,
 		IngressThroughputPerSec:     ingressThroughputPerSec,
 		EgressThroughputPerSec:      egressThroughputPerSec,
-		TotalMaxConnections:         totalMaxConnections,
-		MaxPartitions:               maxPartitions,
+		TotalMaxConnections:         int32(totalMaxConnections),
+		MaxPartitions:               int32(maxPartitions),
 		MaxDataRetentionPeriod:      maxDataRetentionPeriod,
-		MaxConnectionAttemptsPerSec: maxConnectionAttemptsPerSec,
+		MaxConnectionAttemptsPerSec: int32(maxConnectionAttemptsPerSec),
 	}, nil
 }
